// apps/web/src/api/generated/captain/captainDashboardResponseDtoRequiredActionsItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */

export type CaptainDashboardResponseDtoRequiredActionsItem = typeof CaptainDashboardResponseDtoRequiredActionsItem[keyof typeof CaptainDashboardResponseDtoRequiredActionsItem];


export const CaptainDashboardResponseDtoRequiredActionsItem = {
  COMPLETE_PROFILE: 'COMPLETE_PROFILE',
  CREATE_TEAM: 'CREATE_TEAM',
  ADD_ROSTER_PLAYERS: 'ADD_ROSTER_PLAYERS',
  COMPLETE_PAYMENT: 'COMPLETE_PAYMENT',
  WAIT_FOR_APPROVAL: 'WAIT_FOR_APPROVAL',
  CHECK_IN: 'CHECK_IN',
  OPEN_TOURNAMENT_HUB: 'OPEN_TOURNAMENT_HUB',
  VIEW_MATCH: 'VIEW_MATCH',
  NONE: 'NONE',
} as const;